Connection+Crypto: Encrypt at rest data for GET/PUT
Data, both key and value, need to be encrypted at rest. This means that
all values stored in the data structure should be the encrypted version.
An encryption key is generated for the user when they start the server.
Data is encrypted with the key that the server generated, and decrypted
with the key the user provides. If they match, the user gets decrypted
data, otherwise the get garbage values.

Still need to implement DELETE encryption for DELETE.