Main+Connection: Use RwLock instead of Mutex for concurrent reads...
and blocking writes. PUT and DELETE requests are blocking, but GET
requests are not. With the Mutex shared state concurrency, even GET
requests were blocking which is not good and essentially ruins the whole
point of concurrency. Now we can do many reads at once and only lock up
the whole KeyValueStore when doing a modifying operation.

To do this, we had to be specific about what kind of lock, read() or
write(), to acquire in main so we have to move some functionality to
main. And to do that, we have to break apart handle_request into its
pieces and then call each piece individually.